refactor: 캐러셀 내부 동작 로직을 커스텀 훅으로 분리
